frontend/requirements.txt
requests==2.28.1
python-dotenv==0.21.0
pytest==7.1.3
pytest-mock==3.8.2
vcrpy==4.2.1